AdminBaseController moved from components to Controller
- AdminBaseController is jut another Controller and as Component does not really make sense
