+++ packages/framework/src/Controller/Admin/BreadcrumbController.php
@@ -2,7 +2,6 @@
 
 namespace Shopsys\FrameworkBundle\Controller\Admin;
 
-use Shopsys\FrameworkBundle\Component\Controller\AdminBaseController;
 use Shopsys\FrameworkBundle\Model\AdminNavigation\Breadcrumb;
 
 class BreadcrumbController extends AdminBaseController
