+++ packages/framework/src/Controller/Admin/ShopInfoController.php
@@ -3,7 +3,6 @@
 namespace Shopsys\FrameworkBundle\Controller\Admin;
 
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Shopsys\FrameworkBundle\Component\Controller\AdminBaseController;
 use Shopsys\FrameworkBundle\Component\Domain\AdminDomainTabsFacade;
 use Shopsys\FrameworkBundle\Form\Admin\ShopInfo\ShopInfoSettingFormType;
 use Shopsys\FrameworkBundle\Model\ShopInfo\ShopInfoSettingFacade;
